Changes not staged for commit: 	modified:   ../../rsrc/examples/asm 	modified:   ../../rsrc/examples/asm.dSYM/Contents/Resources/DWARF/asm 	modified:   ../../rsrc/examples/compile_champs.sh 	deleted:    asm.dSYM/Contents/Info.plist 	deleted:    ../../../wild/Makefile 	deleted:    ../../../wild/ft_arg.c 	deleted:    ../../../wild/ft_buf.c 	deleted:    ../../../wild/ft_conv_c.c 	deleted:    ../../../wild/ft_conv_d.c 	deleted:    ../../../wild/ft_conv_o.c 	deleted:    ../../../wild/ft_conv_per.c 	deleted:    ../../../wild/ft_conv_s.c 	deleted:    ../../../wild/ft_conv_u.c 	deleted:    ../../../wild/ft_conv_x.c 	deleted:    ../../../wild/ft_fill.c 	deleted:    ../../../wild/ft_handler.c 	deleted:    ../../../wild/ft_itoabase.c 	deleted:    ../../../wild/ft_printf.c 	deleted:    ../../../wild/ft_printf.h 	deleted:    ../../../wild/ft_putwchar.c

+++ Corewar/srcs/asm/write.c
@@ -6,7 +6,7 @@
 /*   By: fmadura <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/12 11:12:05 by fmadura           #+#    #+#             */
-/*   Updated: 2018/07/03 03:32:02 by fmadura          ###   ########.fr       */
+/*   Updated: 2018/07/03 04:22:37 by fmadura          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -65,7 +65,7 @@ int			write_args(int fd, t_ops *ops)
 				write_fill(fd, type ? ops->label[c] : ops->argv[c], 2, 2);
 		}
 		else if (num == 3)
-			write_fill(fd, ops->argv[c], 2, 2);
+			write_fill(fd, (unsigned short)ops->argv[c], 2, 2);
 	}
 	return (1);
 }
